Cleaned up unused files/components
